System18: reorganized configuration files
